requirements.txt
beautifulsoup4
requests
pyaml
lxml